plat-ti: remove console.c
With the simplification of the console/UART code, we're left with only
one function in console.c. Move it to main.c.
