+++ core/arch/arm/plat-ti/sub.mk
@@ -1,4 +1,3 @@
 global-incdirs-y += .
 srcs-y += main.c
-srcs-y += console.c
 srcs-y += sm_platform_handler.c
